refactor(js): move spec-level helpers from JavascriptParser into syntax.js
hasOctalEscape (strict-mode string-escape grammar) and buildLineStarts /
positionAt (ECMAScript LineTerminator semantics — the location decoding
that exists because the lazy parser skips acorn's location tracking)
are parser-level concerns; move them next to the owned parser.
getImportAttributes stays despite decoding syntax's
LEGACY_ASSERT_ATTRIBUTES marker: its `JavascriptParser` static typing is
baked into the generated types.d.ts, noted in place.
