+++ lib/javascript/JavascriptParser.js
@@ -17,7 +17,13 @@ const {
 	parseCommentOptionsInRange
 } = require("../util/magicComment");
 const BasicEvaluatedExpression = require("./BasicEvaluatedExpression");
-const { LEGACY_ASSERT_ATTRIBUTES, WebpackParser } = require("./syntax");
+const {
+	LEGACY_ASSERT_ATTRIBUTES,
+	WebpackParser,
+	buildLineStarts,
+	hasOctalEscape,
+	positionAt
+} = require("./syntax");
 
 /** @typedef {typeof import("acorn").Parser} AcornParser */
 /** @typedef {import("acorn").Options} AcornOptions */
@@ -172,6 +178,8 @@ let parser = /** @type {AcornParser} */ (WebpackParser);
 
 /** @typedef {Record<string, string> & { _isLegacyAssert?: boolean }} ImportAttributes */
 
+// Conceptually pairs with ./syntax's LEGACY_ASSERT_ATTRIBUTES marker, but is
+// pinned here by its public `JavascriptParser.getImportAttributes` typing.
 /**
  * Gets import attributes.
  * @param {ImportDeclaration | ExportNamedDeclaration | ExportAllDeclaration | ImportExpression} node node with assertions
@@ -462,47 +470,6 @@ const isSimpleFunction = (fn) => {
 	return true;
 };
 
-/**
- * Offset of each line's first character. Char-code scan matching acorn's
- * `lineBreak` semantics (CRLF is one break): a regex `exec` loop here
- * allocates a match array per line.
- * @param {string} source source code
- * @returns {number[]} line start offsets
- */
-const buildLineStarts = (source) => {
-	const len = source.length;
-	const lineStarts = [0];
-	for (let i = 0; i < len; i++) {
-		const ch = source.charCodeAt(i);
-		if (ch === 10) {
-			lineStarts.push(i + 1);
-		} else if (ch === 13) {
-			if (source.charCodeAt(i + 1) === 10) i++;
-			lineStarts.push(i + 1);
-		} else if (ch === 0x2028 || ch === 0x2029) {
-			lineStarts.push(i + 1);
-		}
-	}
-	return lineStarts;
-};
-
-/**
- * Binary search for the line containing the offset.
- * @param {number[]} lineStarts line start offsets
- * @param {number} offset source offset
- * @returns {SourcePosition} position (1-based line, 0-based column)
- */
-const positionAt = (lineStarts, offset) => {
-	let lo = 0;
-	let hi = lineStarts.length - 1;
-	while (lo < hi) {
-		const mid = (lo + hi + 1) >>> 1;
-		if (lineStarts[mid] <= offset) lo = mid;
-		else hi = mid - 1;
-	}
-	return { line: lo + 1, column: offset - lineStarts[lo] };
-};
-
 /** @type {ParseOptions} */
 const defaultParserOptions = {
 	sourceType: "module",
@@ -526,33 +493,6 @@ const evaluateAstCaches = new WeakMap();
 const EVALUATE_AST_CACHE_LIMIT = 4096;
 
 const CLASS_NAME = "JavascriptParser";
-
-/**
- * Whether a raw string literal contains a legacy octal (`\47`, `\0` followed by
- * a digit) or non-octal decimal (`\8`, `\9`) escape — all SyntaxErrors in
- * strict mode. Escaped backslashes (`\\`) and `\x` / `\u` escapes are skipped.
- * @param {string} raw raw string literal text, including quotes
- * @returns {boolean} true when a strict-forbidden escape is present
- */
-const hasOctalEscape = (raw) => {
-	for (let i = 0; i < raw.length; i++) {
-		if (raw.charCodeAt(i) !== 92) continue;
-		const next = raw.charCodeAt(i + 1);
-		if (next === 92) {
-			i++;
-			continue;
-		}
-		// `\0` is a valid NUL escape unless a digit follows it.
-		if (next === 48) {
-			const after = raw.charCodeAt(i + 2);
-			if (after >= 48 && after <= 57) return true;
-			i++;
-			continue;
-		}
-		if (next >= 49 && next <= 57) return true;
-	}
-	return false;
-};
 
 class JavascriptParser extends Parser {
 	/**
