+++ lib/javascript/syntax.js
@@ -4177,5 +4177,82 @@ class WebpackParser extends BaseParser {
 	}
 }
 
+/** @typedef {import("../Dependency").SourcePosition} SourcePosition */
+
+/**
+ * Whether a raw string literal contains a legacy octal (`\47`, `\0` followed by
+ * a digit) or non-octal decimal (`\8`, `\9`) escape — all SyntaxErrors in
+ * strict mode. Escaped backslashes (`\\`) and `\x` / `\u` escapes are skipped.
+ * @param {string} raw raw string literal text, including quotes
+ * @returns {boolean} true when a strict-forbidden escape is present
+ */
+const hasOctalEscape = (raw) => {
+	for (let i = 0; i < raw.length; i++) {
+		if (raw.charCodeAt(i) !== 92) continue;
+		const next = raw.charCodeAt(i + 1);
+		if (next === 92) {
+			i++;
+			continue;
+		}
+		// `\0` is a valid NUL escape unless a digit follows it.
+		if (next === 48) {
+			const after = raw.charCodeAt(i + 2);
+			if (after >= 48 && after <= 57) return true;
+			i++;
+			continue;
+		}
+		if (next >= 49 && next <= 57) return true;
+	}
+	return false;
+};
+
+// Location decoding for lazy-mode output: nodes carry only offsets (the parser
+// skips acorn's location tracking), so line/column are derived on demand from
+// these two helpers.
+
+/**
+ * Offset of each line's first character. Char-code scan matching acorn's
+ * `lineBreak` semantics (CRLF is one break): a regex `exec` loop here
+ * allocates a match array per line.
+ * @param {string} source source code
+ * @returns {number[]} line start offsets
+ */
+const buildLineStarts = (source) => {
+	const len = source.length;
+	const lineStarts = [0];
+	for (let i = 0; i < len; i++) {
+		const ch = source.charCodeAt(i);
+		if (ch === 10) {
+			lineStarts.push(i + 1);
+		} else if (ch === 13) {
+			if (source.charCodeAt(i + 1) === 10) i++;
+			lineStarts.push(i + 1);
+		} else if (ch === 0x2028 || ch === 0x2029) {
+			lineStarts.push(i + 1);
+		}
+	}
+	return lineStarts;
+};
+
+/**
+ * Binary search for the line containing the offset.
+ * @param {number[]} lineStarts line start offsets
+ * @param {number} offset source offset
+ * @returns {SourcePosition} position (1-based line, 0-based column)
+ */
+const positionAt = (lineStarts, offset) => {
+	let lo = 0;
+	let hi = lineStarts.length - 1;
+	while (lo < hi) {
+		const mid = (lo + hi + 1) >>> 1;
+		if (lineStarts[mid] <= offset) lo = mid;
+		else hi = mid - 1;
+	}
+	return { line: lo + 1, column: offset - lineStarts[lo] };
+};
+
 module.exports.LEGACY_ASSERT_ATTRIBUTES = LEGACY_ASSERT_ATTRIBUTES;
 module.exports.WebpackParser = WebpackParser;
+module.exports.buildLineStarts = buildLineStarts;
+module.exports.hasOctalEscape = hasOctalEscape;
+module.exports.positionAt = positionAt;
